requirement.txt
fastapi==0.109.2
uvicorn==0.26.0
pydantic[email]==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
motor==3.3.2
python-decouple==3.8
python-multipart==0.0.6
requests==2.31.0
Jinja2==3.1.3
beanie==1.25.0